OT104-81 changed service export from cunstom hook to object. Changed functions name to suite the team.
