+++ src/Services/slidesService.test.js
@@ -1,23 +1,21 @@
-import { useSlides } from "./slides";
-
-const slides = useSlides();
+import slides from "./slidesService";
 
 test("useSlides return an object", () => {
   expect(typeof slides).toBe("object");
 });
 
 test("slides should contains a function getSlides", () => {
-  expect(typeof slides.getSlides).toBe("function");
+  expect(typeof slides.get).toBe("function");
 });
 test("slides should contains a function getSlide", () => {
-  expect(typeof slides.getSlide).toBe("function");
+  expect(typeof slides.getAll).toBe("function");
 });
 test("slides should contains a function createSlide", () => {
-  expect(typeof slides.createSlide).toBe("function");
+  expect(typeof slides.create).toBe("function");
 });
 test("slides should contains a function updateSlide", () => {
-  expect(typeof slides.updateSlide).toBe("function");
+  expect(typeof slides.update).toBe("function");
 });
 test("slides should contains a function deleteSlide", () => {
-  expect(typeof slides.deleteSlide).toBe("function");
+  expect(typeof slides.delete).toBe("function");
 });
